+++ integration-tests/sanity-functional/src/test/java/com/cognifide/aet/sanity/functional/HomePageTilesTest.java
@@ -29,9 +29,9 @@ import org.junit.runner.RunWith;
 @Modules(GuiceModule.class)
 public class HomePageTilesTest {
 
-  private static final int TESTS = 149;
+  private static final int TESTS = 157;
 
-  private static final int EXPECTED_TESTS_SUCCESS = 85;
+  private static final int EXPECTED_TESTS_SUCCESS = 93;
 
   private static final int EXPECTED_TESTS_CONDITIONALLY_PASSED = 11;