+++ report/src/main/webapp/app/app.config.js
@@ -61,6 +61,7 @@ require.config({
     //sidepanel
     'sidepanelDirective': 'layout/sidepanel/sidepanel.directive',
     'sidepanelStatusFilterDirective': 'layout/sidepanel/sidepanelStatusFilter.directive',
+    'sidepanelOrderDirective': 'layout/sidepanel/sidepanelOrder.directive',
     'sidepanelSearchDirective': 'layout/sidepanel/sidepanelSearch.directive',
     'sidepanelToggleLinkDirective': 'layout/sidepanel/toggleLink.directive',
     'sidepanelSaveChangesDirective': 'layout/sidepanel/saveChanges.directive',
@@ -77,6 +78,7 @@ require.config({
     //modals
     'unsavedChangesModalController': 'layout/modal/unsavedChanges/unsavedChangesModal.controller',
     'noteModalController': 'layout/modal/note/noteModal.controller',
+    'accessibilityModalController': 'layout/modal/accessibility/accessibilityModal.controller',
     'historyModalController': 'layout/modal/history/historyModal.controller'
   },
   shim: {