+++ report/src/main/webapp/app/layout/main/suite/mainView.suite.controller.js
@@ -31,7 +31,20 @@ define([], function () {
      ***************************************/
 
     function updateSuiteView() {
-      vm.tests = metadataAccessService.getTests();
+      var tests = metadataAccessService.getTests();
+      vm.testWrappers = [];
+      _.forEach(tests, function (test) {
+        var urlStats = {};
+        _.forEach(test.urls, function (url) {
+          var status = url.getStatus();
+          urlStats[status] = urlStats[status] ? urlStats[status] + 1 : 1;
+        });
+        vm.testWrappers.push({
+          name: test.name,
+          test: test,
+          stats: urlStats
+        });
+      });
     }
   }
 });